Fixed a bug in handling the counter axis.
In the mess-program the counter array was 1 element too long
(n_shots+1). The last element remained 0. The whole array was then
written to the file. Since we used the last n_shots of the imported
data, it seemed as if the last element was 0. What really happened is
that we used the range [1:n_shots+1].