Fixed a mistake in the axis when importing data binned in VB6
We now implemented binning in VB6. However, the time axis for the
non-rephasing diagram was reversed.
The obvious place to fix it is during importing, but I decided against
it because it will influence the bin-counts, axes etc. and we want to
do measurements tomorrow. At this place the issue and the solution are
clear: reverse the axis.